+++ XUnitTest.XCode/Code/entity_log_normal_biz.cs
@@ -156,7 +156,7 @@ public partial class Log : Entity<Log>
     /// <param name="category">类别</param>
     /// <param name="linkId">链接</param>
     /// <returns>实体列表</returns>
-    public static IList<Log> FindAllByCategoryAndLinkID(String category, Int32 linkId)
+    public static IList<Log> FindAllByCategoryAndLinkID(String category, Int64 linkId)
     {
         // 实体缓存
         if (Meta.Session.Count < 1000) return Meta.Cache.FindAll(e => e.Category.EqualIgnoreCase(category) && e.LinkID == linkId);
@@ -189,7 +189,7 @@ public partial class Log : Entity<Log>
     /// <param name="key">关键字</param>
     /// <param name="page">分页参数信息。可携带统计和数据权限扩展查询等信息</param>
     /// <returns>实体列表</returns>
-    public static IList<Log> Search(String category, String action, Int32 linkId, Int32 createUserId, DateTime start, DateTime end, String key, PageParameter page)
+    public static IList<Log> Search(String category, String action, Int64 linkId, Int32 createUserId, DateTime start, DateTime end, String key, PageParameter page)
     {
         var exp = new WhereExpression();
